Add a PostingSpec class
There is a separate Posting and PostingSpec class. This introduces a lot
of verbosity, but at least this way we're always sure if the amount is
present, and it'll avoid a tons of extra checks in the engine regarding
if the transaction has been resolved or not.

Eventually, I'd like to introduce a TransactionSpec as well, so that
there is no confusion.