devenv: Remove cocapods + ruby
This is only required in mac, and it fails to install in the CI which
is a Linux based env

The whole promise of only needing 1 env across different platforms with
nix is turning out to be quite a lie.

Ideally, I would just conditionally skip this on macos, but my nix
skills are too low to do that.