Remove custom DateRange class
It's getting confusing between having a DateTimeRange class provided by
Flutter, and this DateRange class having both the start + end Date
optional.

It seems better to design an API where I can explicitly specify if the
start or endDate is nullable or not, and just treat them like two
parameters.