Transformers: We now operate on a Map<String, String>
Instead of a List<String>. Since most CSVs have headers, this way we can
select a field based on the header name and not rely on the index
position on the List, as that is far more fragile.

For the cases when we don't have headers, then we can always just make
them up, as has been done for the tests.