Try to write transformers for the Wise currency conversion
This required writing a Transformer that added two numbers. I don't
think this is a good way to go about it, as building a
TransformerBuilder for this is challenging.

Instead it would be better if Beany added support for expressions like
Beancount has, so that the transformer builder can see the invidual
parts of the calculation. "x + y" for example.

Adding this is going to be a huge challenge as the Beany core systems
haven't been built for expressions. Bah. I guess I'll have to a
DecimalSpec class or something.