Remove all wise converter code
Wise is now exactly the same as all the other banks and just provides a
standard csv. I'd prefer to use the Transformer framework to train a
transformer for it, or build the transformer manually instead of this
hand written code.

Also, the CSV format seems to now be a bit different anyway.