importer: TranformerBuilder: Always add the FixedCurrency
We were only adding it if there were no currencies transformers found,
but that's problematic as sometimes it's only there in some lines, and
not all of them.

And then when figuring out the common transformers, we don't land up
getting the fewest, and need to rely on the decision tree. It's easier
if there is just one transformer.